Split stats.swift into two files Common.swift and Distributions.swift
